fix: fix expEarned variable type to double
BREAKING CHANGE: Ln 33 and Ln 43, expEarned was an int variable type

+++ my_farm/Seeds.java
@@ -30,7 +30,7 @@ public class Seeds {
       this.produceMax = produceMax;
    }
 
-   public Seeds(String seedName, String cropType, int harvestTime, int waterNeed, int waterBonus, int fertilizerNeed, int fertilizerBonus, int seedCost, int basePrice, int expEarned, int produceMin) {
+   public Seeds(String seedName, String cropType, int harvestTime, int waterNeed, int waterBonus, int fertilizerNeed, int fertilizerBonus, int seedCost, int basePrice, double expEarned, int produceMin) {
       this.seedName = seedName;
       this.cropType = cropType;
       this.harvestTime = harvestTime;
